+++ server/TWS Foundation/TWS Security/Entities/Accounts/IAccountsDepot.cs
@@ -3,5 +3,5 @@
 namespace TWS_Security.Entities.Accounts;
 public interface IAccountsDepot 
     : IDepot<Account> {
-    public Task<Permit[]> GetPermits(int Account);
+    public Task<Permit[]> GetPermits(long Account);
 }
